+++ src/components/field/field-select.tsx
@@ -1,6 +1,5 @@
-import { CaretUpDownIcon } from '@phosphor-icons/react'
 import { useFieldContext } from '../form/form-context'
-import { Select, SelectIcon, SelectList, SelectPopup, SelectTrigger, SelectValue } from '../select'
+import { Select } from '../select'
 
 import type { FieldSelectProps } from './field.types'
 
@@ -14,15 +13,7 @@ export const FieldSelect = ({ children, ...props }: FieldSelectProps) => {
       onValueChange={(value) => field.setValue(value)}
       {...props}
     >
-      <SelectTrigger>
-        <SelectValue />
-        <SelectIcon>
-          <CaretUpDownIcon weight="bold" />
-        </SelectIcon>
-      </SelectTrigger>
-      <SelectPopup>
-        <SelectList>{children}</SelectList>
-      </SelectPopup>
+      {children}
     </Select>
   )
 }