+++ src/components/select/select.types.ts
@@ -3,7 +3,7 @@ import { cva, type VariantProps } from 'class-variance-authority'
 import type { Select as BaseSelect } from '@base-ui/react/select'
 
 export const selectVariants = cva(
-  'flex min-w-[calc(var(--dawn-spacing-3xl)+8rem)] items-center justify-between outline outline-transparent transition-all hover:cursor-pointer aria-invalid:bg-error-container aria-invalid:text-error-on-container aria-invalid:outline-error-border data-[disabled=true]:cursor-not-allowed data-[disabled=true]:opacity-50',
+  'flex items-center justify-between outline outline-transparent transition-all hover:cursor-pointer aria-invalid:bg-error-container aria-invalid:text-error-on-container aria-invalid:outline-error-border data-[disabled=true]:cursor-not-allowed data-[disabled=true]:opacity-50 [&_[data-value]]:min-w-0 [&_[data-value]]:flex-1',
   {
     variants: {
       variant: {
@@ -13,11 +13,11 @@ export const selectVariants = cva(
       },
       size: {
         small:
-          'h-lg rounded-lg px-xs style-text-default--1 [&_[data-value]]:gap-3xs [&>svg]:size-xs',
+          'h-lg gap-md rounded-lg px-xs style-text-default--1 [&_[data-value]]:gap-3xs [&_svg]:size-xs',
         medium:
-          'h-xl rounded-xl px-sm style-text-default-0 [&_[data-value]]:gap-2xs [&>svg]:size-sm',
+          'h-xl gap-lg rounded-xl px-sm style-text-default-0 [&_[data-value]]:gap-2xs [&_svg]:size-sm',
         large:
-          'h-2xl rounded-2xl px-md style-text-default-1 [&_[data-value]]:gap-xs [&>svg]:size-md',
+          'h-2xl gap-xl rounded-2xl px-md style-text-default-1 [&_[data-value]]:gap-xs [&_svg]:size-md',
       },
     },
     defaultVariants: {
@@ -37,3 +37,8 @@ export type SelectItemProps = React.ComponentProps<typeof BaseSelect.Item>
 export type SelectListProps = React.ComponentProps<typeof BaseSelect.List>
 export type SelectGroupProps = React.ComponentProps<typeof BaseSelect.Group>
 export type SelectGroupLabelProps = React.ComponentProps<typeof BaseSelect.GroupLabel>
+export type SelectFilter = {
+  id: number
+  label: string
+  value: string
+}