+++ src/components/table/table-row-list.tsx
@@ -1,15 +1,33 @@
 import { flexRender } from '@tanstack/react-table'
+import { useTable } from './table'
 import { cn } from '@/utils/cn'
 
 import type { TableRowListProps } from './table.types'
 
-export const TableRowList = ({ row, className, children, ref, ...props }: TableRowListProps) => {
+export const TableRowList = <TData,>({
+  row,
+  className,
+  children,
+  ref,
+  ...props
+}: TableRowListProps<TData>) => {
+  const { table } = useTable()
+
+  const isSelected = table.getSelectedRowModel().flatRows.some((r) => r.id === row.id)
+
   return (
-    <tr className={cn('', className)} ref={ref} {...props}>
+    <tr
+      className={cn('transition-colors', className)}
+      ref={ref}
+      style={{
+        backgroundColor: isSelected ? 'var(--color-neutral-container-high)' : 'transparent',
+      }}
+      {...props}
+    >
       {row.getVisibleCells().map((cell) => {
         return (
           <td
-            className="p-sm"
+            className="px-sm py-xs first:rounded-l-full last:rounded-r-full"
             key={cell.id}
             style={{
               width: cell.column.getSize(),