+++ src/components/table/table.tsx
@@ -1,29 +1,46 @@
 import * as React from 'react'
+import { cn } from '@/utils/cn'
 
-import type { TableProps } from './table.types'
+import type { FilterType, TableDataSet, TableProps, TableView } from './table.types'
+import type { RowData } from '@tanstack/react-table'
 
-type TableContextType = TableProps & {
+type TableContextType<TData, TValue> = {
+  table: TableProps<TData, TValue>['table']
+  dataSets?: TableDataSet<TData, TValue>[]
+  currentDataSet?: TableDataSet<TData, TValue>
+  setCurrentDataSet?: React.Dispatch<React.SetStateAction<TableDataSet<TData, TValue>>>
   view: TableView
   setView: React.Dispatch<React.SetStateAction<TableView>>
 }
 
-const TableContext = React.createContext<TableContextType | null>(null)
+declare module '@tanstack/react-table' {
+  interface ColumnMeta<TData extends RowData, TValue> {
+    filterType?: FilterType
+  }
+}
 
-export type TableView = 'grid' | 'list'
+const TableContext = React.createContext<TableContextType<any, any> | null>(null)
 
-export const Table = ({
+export const Table = <TData, TValue>({
+  currentDataSet,
+  setCurrentDataSet,
+  dataSets,
   defaultView = 'list',
   table,
   className,
   children,
   ref,
   ...props
-}: TableProps) => {
+}: TableProps<TData, TValue>) => {
   const [view, setView] = React.useState<TableView>(defaultView)
 
   return (
-    <TableContext.Provider value={{ view, setView, table, className, children, ref, ...props }}>
-      <div className="flex flex-col gap-sm">{children}</div>
+    <TableContext.Provider
+      value={{ dataSets, currentDataSet, setCurrentDataSet, table, view, setView }}
+    >
+      <div className={cn('flex flex-col gap-xs', className)} {...props} ref={ref}>
+        {children}
+      </div>
     </TableContext.Provider>
   )
 }